chore: fixes after junit and keys module refactoring
