+++ edc-extensions/iatp/tx-iatp-sts-dim/build.gradle.kts
@@ -29,4 +29,5 @@ dependencies {
     implementation(project(":core:core-utils"))
 
     testImplementation(libs.edc.junit)
+    testImplementation(testFixtures(libs.edc.lib.http))
 }
